Common: Purge pxStreams and some other wx string routines
