+++ common/CMakeLists.txt
@@ -27,7 +27,6 @@ target_sources(common PRIVATE
 	PrecompiledHeader.cpp
 	Perf.cpp
 	ProgressCallback.cpp
-	pxStreams.cpp
 	pxTranslate.cpp
 	RwMutex.cpp
 	Semaphore.cpp
@@ -84,7 +83,6 @@ target_sources(common PRIVATE
 	PrecompiledHeader.h
 	ProgressCallback.h
 	pxForwardDefs.h
-	pxStreams.h
 	RedtapeWindows.h
 	RwMutex.h
 	SafeArray.h
@@ -98,7 +96,6 @@ target_sources(common PRIVATE
 	Threading.h
 	TraceLog.h
 	WindowInfo.h
-	wxBaseTools.h
 	emitter/cpudetect_internal.h
 	emitter/implement/dwshift.h
 	emitter/implement/group1.h
