+++ common/common.vcxproj
@@ -61,7 +61,6 @@
     <ClCompile Include="IniInterface.cpp" />
     <ClCompile Include="MD5Digest.cpp" />
     <ClCompile Include="ProgressCallback.cpp" />
-    <ClCompile Include="pxStreams.cpp" />
     <ClCompile Include="pxTranslate.cpp" />
     <ClCompile Include="StringUtil.cpp" />
     <ClCompile Include="SettingsWrapper.cpp" />
@@ -161,7 +160,6 @@
     <ClInclude Include="Vulkan\Texture.h" />
     <ClInclude Include="Vulkan\Util.h" />
     <ClInclude Include="WindowInfo.h" />
-    <ClInclude Include="wxBaseTools.h" />
     <ClInclude Include="Threading.h" />
     <ClInclude Include="PersistentThread.h" />
     <ClInclude Include="RwMutex.h" />
