+++ pcsx2/PAD/Windows/Global.h
@@ -20,7 +20,6 @@
 #include <wx/intl.h>
 #include <wx/log.h>
 #include <wx/filename.h>
-#include "common/pxStreams.h"
 #include "common/Console.h"
 #include <stdio.h>
 #include <assert.h>
