+++ pcsx2/SPU2/spu2.cpp
@@ -28,7 +28,6 @@
 #include "Host/Dialogs.h"
 #endif
 #include "R3000A.h"
-#include "common/pxStreams.h"
 
 using namespace Threading;
